wificar/requirements.txt
pygame==1.9.6
Adafruit_ADS1x15==1.0.2
Adafruit_GPIO==1.0.3
Adafruit_PCA9685==1.0.1
Adafruit_SSD1306==1.6.2
Pillow==6.1.0
pyserial==3.4
